main: handle query selectors that are only usable for single packages correctly
--print-requires, --print-requires-private, --print-provides, --modversion, --print-variable and --print-variables
are inquiring information about a package, not about a dependency graph.

in such cases, we should not try to solve the problem set to determine if it is satisfiable, as any invocation
of these commands are trying to inquire about a specific package.  instead, just solve the problem set for a single
level (so we can load the package itself) which is far more useful behaviour when scanning .pc files for their
dependency metadata.

as an aside, this allows us to successfully scan .pc files which are otherwise unsatisfiable.