pkg: make pkgconf_compare_version consistent
The code taken from rpmvercmp in pkg-config returns -1 if a is less than 
b, 0 if a is equal to b, and 1 if a is greater than b. This matches the  
expectations of the comparison operators that use this function.         
However, the tilde handling, the NULL handling, and the docstring all do 
the opposite.                                                            
                                                                         
This fixes the tilde handling, the NULL handling, and the docstring to   
match the behavior of the rpmvercmp code and the expectations of the     
comparison operators.