libpkgconf: parser: gracefully handle pkgconf_fgetline failures
With the move to dynamic buffering of pkg-config files, it is possible for
pkgconf_fgetline to leave the buffer in an unwritten state, where a buffer
will not be allocated.

We must gracefully handle this situation, so we bail early and re-enter the
loop to try reading the buffer again.  If it fails again, then we finish
parsing the already read pkg-config data.

Related: https://github.com/pkgconf/pkgconf/issues/384
Signed-off-by: Ariadne Conill <ariadne@ariadne.space>