libpkgconf: fragment: fix fragment length estimation
The switch from flat fragment lists to fragment trees made the fragment length
estimation more complicated.  Originally, the code when migrated just probed the
length of child fragments directly, but this only works for fragment trees where
there is only one level of children.

In some cases, there may be more children at a lower depth that were not properly
accounted for in the length estimation, which led to truncated buffers being
generated.

Fixes: https://github.com/pkgconf/pkgconf/issues/383
Signed-off-by: Ariadne Conill <ariadne@ariadne.space>