libpkgconf: path: add pkgconf_path_prepend_list
Previously when processing --with-path entries we were adding to the tail
of the search list.  In reality, we expect --with-path entries to be treated
in an equivalent way to PKG_CONFIG_PATH.

Accordingly, add a variant of pkgconf_path_copy_list which prepends instead.

Related: https://github.com/pkgconf/pkgconf/issues/400
Signed-off-by: Ariadne Conill <ariadne@ariadne.space>