man/pkgconf.1: correctly describe the module version check options
The options --atleast-version, --exact-version, and --max-version
short-circuit to both success and failure, which implies that the
result depends on the the order of the arguments and that a simple
description is not possible.  In particular, it is not possible
to describe the behaviour with any sentence of the form
"(succeed|fail) if (any|each) argument satisfies (condition)".